+++ src/main/java/org/dcsa/standards/specifications/standards/core/v103/types/FacilityTypeCode.java
@@ -1,0 +1,30 @@
+package org.dcsa.standards.specifications.standards.core.v103.types;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.dcsa.standards.specifications.generator.EnumBase;
+
+@Getter
+@Schema(
+    type = "string",
+    maxLength = 10,
+    example = "RAMP",
+    description =
+"""
+Code identifying a specific type of facility
+""")
+@AllArgsConstructor
+public enum FacilityTypeCode implements EnumBase {
+  BORD("Border"),
+  CLOC("Customer location"),
+  COFS("Container freight station"),
+  OFFD("Off-dock storage"),
+  DEPO("Depot"),
+  INTE("Inland terminal"),
+  POTE("Port terminal"),
+  RAMP("Ramp"),
+  WAYP("Waypoint");
+
+  private final String valueDescription;
+}