+++ src/main/java/org/dcsa/standards/specifications/standards/core/v103/types/UNLocationCode.java
@@ -1,0 +1,18 @@
+package org.dcsa.standards.specifications.standards.core.v103.types;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+
+@Schema(
+    type = "string",
+    maxLength = 5,
+    example = "NLAMS",
+    description =
+"""
+The UN Location code specifying where the place is located. The pattern used must be
+
+- 2 characters for the country code using [ISO 3166-1 alpha-2](https://www.iso.org/obp/ui/#iso:pub:PUB500001:en)
+- 3 characters to code a location within that country. Letters A-Z and numbers from 2-9 can be used
+
+More info can be found here: [UN/LOCODE](https://unece.org/trade/cefact/UNLOCODE-Download)
+""")
+public class UNLocationCode {}