+++ src/main/java/org/dcsa/standards/specifications/standards/tnt/v300/model/EquipmentDetails.java
@@ -3,9 +3,9 @@ package org.dcsa.standards.specifications.standards.tnt.v300.model;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import lombok.Data;
-import org.dcsa.standards.specifications.standards.core.v101.model.Seal;
-import org.dcsa.standards.specifications.standards.core.v101.types.EquipmentReference;
-import org.dcsa.standards.specifications.standards.core.v101.types.IsoEquipmentCode;
+import org.dcsa.standards.specifications.standards.core.v103.model.Seal;
+import org.dcsa.standards.specifications.standards.core.v103.types.EquipmentReference;
+import org.dcsa.standards.specifications.standards.core.v103.types.IsoEquipmentCode;
 import org.dcsa.standards.specifications.standards.tnt.v300.types.EmptyIndicatorCode;
 import org.dcsa.standards.specifications.standards.tnt.v300.types.MovementTypeCode;