+++ src/main/java/org/dcsa/standards/specifications/standards/tnt/v300/model/EventRouting.java
@@ -3,7 +3,7 @@ package org.dcsa.standards.specifications.standards.tnt.v300.model;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.List;
 import lombok.Data;
-import org.dcsa.standards.specifications.standards.core.v101.model.Party;
+import org.dcsa.standards.specifications.standards.core.v103.model.Party;
 
 @Schema(
     description =