+++ src/main/java/org/dcsa/standards/specifications/standards/tnt/v300/model/TransportDetails.java
@@ -2,7 +2,7 @@ package org.dcsa.standards.specifications.standards.tnt.v300.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
-import org.dcsa.standards.specifications.standards.core.v101.types.DelayReasonCode;
+import org.dcsa.standards.specifications.standards.core.v103.types.DelayReasonCode;
 
 import java.util.List;